Major refactor in preparation for asyncio
Factor out the event loop core from the ``EventListener``
and into its own module. Move the listener into a ``handlers`` module.
