+++ switchy/__init__.py
@@ -9,7 +9,9 @@ Licensed under the MPL 2.0 license (see `LICENSE` file)
 from os import path
 from . import apps
 from .utils import get_logger, ESLError
-from .observe import EventListener, Client, get_listener
+from .observe import Client
+from .async import get_event_loop
+from .handlers import get_listener
 from .apps.call_gen import get_originator
 from .distribute import SlavePool, MultiEval
 from .marks import event_callback, handler
@@ -18,7 +20,7 @@ from .sync import sync_caller
 from .serve import Service
 
 __package__ = 'switchy'
-__version__ = '0.1.alpha'
+__version__ = '0.1.0.alpha1'
 __author__ = ('Sangoma Technologies', 'qa@eng.sangoma.com')
