+++ switchy/distribute.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 """
-Manage pools of freeswitch slaves
+Manage clients over a cluster of FreeSWITCH processes.
 """
 from itertools import cycle
 from operator import add
